+++ test/unit/fuzz-tools.test.js
@@ -245,6 +245,7 @@ import * as orthospineV252 from '../../lib/orthospine-v252.js';
 import * as radmeasureV253 from '../../lib/radmeasure-v253.js';
 import * as enturopsychV254 from '../../lib/enturopsych-v254.js';
 import * as riskscoresV255 from '../../lib/riskscores-v255.js';
+import * as rheumcritV256 from '../../lib/rheumcrit-v256.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -434,6 +435,7 @@ const MODULES = {
   'radmeasure-v253.js': radmeasureV253,
   'enturopsych-v254.js': enturopsychV254,
   'riskscores-v255.js': riskscoresV255,
+  'rheumcrit-v256.js': rheumcritV256,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];